Remove duplicate database code
Import customer database instead of replicating.
